+++ Content.Shared/Containers/ContainerFillSystem.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using System.Numerics;
 using Content.Shared.EntityTable;
-using Content.Shared.Storage.EntitySystems;
+using Content.Shared.Item;
 using Robust.Shared.Containers;
 using Robust.Shared.Map;
 
@@ -12,7 +12,7 @@ public sealed partial class ContainerFillSystem : EntitySystem
     [Dependency] private SharedContainerSystem _containerSystem = default!;
     [Dependency] private EntityTableSystem _entityTable = default!;
     [Dependency] private SharedTransformSystem _transform = default!;
-    [Dependency] private SharedStorageSystem _storageSys = default!;
+    [Dependency] private SharedItemSystem _itemSys = default!;
 
     public override void Initialize()
     {
@@ -75,7 +75,7 @@ public sealed partial class ContainerFillSystem : EntitySystem
             if (ent.Comp.Sort)
             {
                 // Reverse order since we want to insert larger items first, and the list is sorted smallest to largest.
-                spawns.Sort((a, b) => _storageSys.CompareSize(b, a));
+                spawns.Sort((a, b) => _itemSys.CompareSize(b, a));
             }
 
             foreach (var proto in spawns)
