+++ Content.Shared/Item/SharedItemSystem.cs
@@ -18,6 +18,7 @@ public abstract partial class SharedItemSystem : EntitySystem
     [Dependency] private IPrototypeManager _prototype = default!;
     [Dependency] private SharedHandsSystem _handsSystem = default!;
     [Dependency] protected SharedContainerSystem Container = default!;
+    [Dependency] private IComponentFactory _compFactory = default!;
 
     public override void Initialize()
     {
@@ -270,5 +271,26 @@ public abstract partial class SharedItemSystem : EntitySystem
                 SetSize(uid, (ProtoId<ItemSizePrototype>) itemToggleSize.DeactivatedSize, item);
             }
         }
+    }
+
+    /// <summary>
+    /// Sorts two protos by <see cref="ItemComponent"/> size, from smallest to largest.
+    /// </summary>
+    /// <param name="a">The first proto.</param>
+    /// <param name="b">The second proto.</param>
+    /// <returns> Less than 0 if a is smaller, greater than 0 if a is larger,
+    /// 0 if they are the same or either proto doesn't have an <see cref="ItemComponent"/>.</returns>
+    [PublicAPI]
+    public int CompareSize(EntProtoId a, EntProtoId b)
+    {
+        var protoA = _prototype.Index(a);
+        var protoB = _prototype.Index(b);
+        if (!protoA.TryGetComponent<ItemComponent>(out var compA, _compFactory) ||
+            !protoB.TryGetComponent<ItemComponent>(out var compB, _compFactory))
+        {
+            return 0;
+        }
+
+        return _prototype.Index(compA.Size).CompareTo(_prototype.Index(compB.Size));
     }
 }
